Address review: pin MSDF cache types in guard; scale advance in double
msdf_text_api_guard.cpp: also pin the serialization-load-bearing types
(kerning_key_t == uint64_t, the kerning_units/glyphs map types, and rgba's
element type). font_renderer writes these to the on-disk font cache, so a width
or shape change upstream would silently corrupt the format; the canary now fails
to compile instead of relying solely on the runtime cache-version bump to reject.

font_renderer.cpp: compute draw_scale_for in double to mirror the library's
draw_scaling_for(), so the scaled monospace advance matches scaled_glyph()'s
advance_x; cast the product to float at the call site.

+++ src/core/font_renderer.cpp
@@ -224,13 +224,15 @@ bool validate_cached_glyph(const msdf_glyph_t& g)
 }
 
 // draw_scale matching the library's draw_scaling_for(): draw_pixel_height /
-// font-unit ascender. Used to project a font-unit advance to output pixels.
-float draw_scale_for(const msdf_atlas_t& atlas, int draw_pixel_height)
+// font-unit ascender, computed in double like the library so a font-unit advance
+// scaled here matches scaled_glyph().advance_x. Keep in sync with the library if
+// it ever changes how advances scale.
+double draw_scale_for(const msdf_atlas_t& atlas, int draw_pixel_height)
 {
-    const float ascender = atlas.font_metrics_units.ascender;
-    return (ascender > 0.f)
-        ? static_cast<float>(draw_pixel_height) / ascender
-        : 0.f;
+    const double ascender = atlas.font_metrics_units.ascender;
+    return (ascender > 0.0)
+        ? static_cast<double>(draw_pixel_height) / ascender
+        : 0.0;
 }
 
 void add_text_to_vectors(
@@ -883,8 +885,9 @@ float Font_renderer::monospace_advance_px() const
     if (!atlas) {
         return 0.f;
     }
-    return atlas->zero_advance_units *
-        draw_scale_for(*atlas, m_impl->current_draw_pixel_height());
+    return static_cast<float>(
+        atlas->zero_advance_units *
+        draw_scale_for(*atlas, m_impl->current_draw_pixel_height()));
 }
 
 bool Font_renderer::monospace_advance_is_reliable() const
